me me big boy

+++ src/com/lapissea/opengl/program/game/terrain/IHeightMapProvider.java
@@ -2,6 +2,6 @@ package com.lapissea.opengl.program.game.terrain;
 
 public interface IHeightMapProvider{
 	
-	public float getHeightAt(double x, double y);
+	public double getHeightAt(double x, double y);
 	
 }
